+++ src/render_player.cpp
@@ -11,8 +11,10 @@ void render_player()
     SDL_Point offset, tex_size;
     auto texture = get_player_sprite(&offset, &tex_size);
     SDL_Rect sprite_area {
-        player.pos.x + offset.x - camera.x,
-        player.pos.y + offset.y - camera.y,
+        static_cast<int>(player.pos.x)
+            + offset.x - camera.x,
+        static_cast<int>(player.pos.y)
+            + offset.y - camera.y,
         tex_size.x,
         tex_size.y
     };
@@ -30,8 +32,8 @@ void render_player()
 
     if(SHOW_DEBUG_GRID) { // Player hitbox
         SDL_Rect player_area {
-            player.pos.x - camera.x,
-            player.pos.y - camera.y,
+            static_cast<int>(player.pos.x) - camera.x,
+            static_cast<int>(player.pos.y) - camera.y,
             player.size.x,
             player.size.y
         };
